+++ test/e2e/hook_test.go
@@ -149,3 +149,33 @@ func TestHookDeletePolicyHookFailedHookExit1(t *testing.T) {
 		Expect(OperationPhaseIs(OperationFailed)).
 		Expect(NotPod(func(p v1.Pod) bool { return p.Name == "hook" }))
 }
+
+func TestFixingDegradedApp(t *testing.T) {
+	Given(t).
+		Path("sync-waves").
+		When().
+		Create().
+		Sync().
+		Then().
+		Expect(ResourceSyncStatusIs("pod-1", SyncStatusCodeSynced)).
+		Expect(ResourceHealthIs("pod-1", HealthStatusHealthy)).
+		Expect(ResourceSyncStatusIs("pod-2", SyncStatusCodeOutOfSync)).
+		Expect(ResourceHealthIs("pod-2", HealthStatusMissing)).
+		Expect(SyncStatusIs(SyncStatusCodeOutOfSync)).
+		Expect(HealthIs(HealthStatusMissing)).
+		When().
+		TerminateOp().
+		Then().
+		Expect(OperationPhaseIs(OperationFailed)).
+		When().
+		Patch("pod-1.yaml", `[{"op": "replace", "path": "/spec/containers/0/image", "value": "nginx"}]`).
+		Sync().
+		Then().
+		Expect(ResourceSyncStatusIs("pod-1", SyncStatusCodeSynced)).
+		Expect(ResourceHealthIs("pod-1", HealthStatusHealthy)).
+		Expect(ResourceSyncStatusIs("pod-2", SyncStatusCodeSynced)).
+		Expect(ResourceHealthIs("pod-2", HealthStatusHealthy)).
+		Expect(SyncStatusIs(SyncStatusCodeSynced)).
+		Expect(HealthIs(HealthStatusHealthy)).
+		Expect(OperationPhaseIs(OperationSucceeded))
+}
